Give a hint about the list of tutorial pages under the Prev/Next buttons
It didn't occur to me to click on the menu icon in the upper left to get a list of tutorial pages,
because I had clicked an identical menu icon to choose which version of the tutorial to view.
There is no indication that the functionality of that icon has changed.